Agentic_AI/cognitive_guardian/requirements.txt
# Core Python
python-dotenv>=1.0.0

# Agent Framework
langgraph>=0.0.40
langchain>=0.1.0

# DeepSeek (OpenAI compatible client)
openai>=1.12.0

# UI Frameworks
gradio>=4.20.0
streamlit>=1.32.0
streamlit-extras>=0.4.0

# Data / Visualization
plotly>=5.18.0

# YouTube API
google-api-python-client>=2.115.0

# Semantic Matching
scikit-learn>=1.4.0

# Utilities
pydantic>=2.5.0
requests>=2.31.0
